Add new library for participants to modify during the Kata.
