+++ clients/cpp/CMakeLists.txt
@@ -17,29 +17,7 @@ include(cmake/FindGoogleTest.cmake)
 include(cmake/FindNlohmannJson.cmake)
 include(cmake/FindBoost.cmake)
 
-#Functional library to be modified by trainees
-add_library(Extreme-Carpaccio-Order-Management-shared SHARED)
-
-target_compile_definitions(Extreme-Carpaccio-Order-Management-shared PRIVATE EXTREME_CARPACCIO_ORDER_MANAGEMENT_EXPORTS)
-
-target_sources(Extreme-Carpaccio-Order-Management-shared
-    PRIVATE
-        ${CMAKE_CURRENT_SOURCE_DIR}/src/extreme_carpaccio_order_management/OrderParsing.cpp
-		${CMAKE_CURRENT_SOURCE_DIR}/src/extreme_carpaccio_order_management/TotalAmount.cpp
-    PUBLIC
-        ${CMAKE_CURRENT_SOURCE_DIR}/include/extreme_carpaccio_order_management/OrderParsing.hpp
-		${CMAKE_CURRENT_SOURCE_DIR}/include/extreme_carpaccio_order_management/TotalAmount.hpp
-        ${CMAKE_CURRENT_SOURCE_DIR}/include/extreme_carpaccio_order_management/config.hpp
-)
-
-target_include_directories(Extreme-Carpaccio-Order-Management-shared
-    PRIVATE
-        ${CMAKE_CURRENT_SOURCE_DIR}/src
-    PUBLIC
-        ${CMAKE_CURRENT_SOURCE_DIR}/include
-)
-
-target_link_libraries(Extreme-Carpaccio-Order-Management-shared PUBLIC Boost::boost PRIVATE nlohmann_json)
+add_subdirectory(OrderManagement)
 
 #Shared library
 add_library(Extreme-Carpaccio-Client-shared SHARED)
@@ -49,12 +27,10 @@ target_compile_definitions(Extreme-Carpaccio-Client-shared PRIVATE EXTREME_CARPA
 target_sources(Extreme-Carpaccio-Client-shared
     PRIVATE
         ${CMAKE_CURRENT_SOURCE_DIR}/src/extreme_carpaccio_client/ExtremeCarpaccioClient.cpp
-		${CMAKE_CURRENT_SOURCE_DIR}/src/extreme_carpaccio_client/Order.cpp
 		${CMAKE_CURRENT_SOURCE_DIR}/src/extreme_carpaccio_client/HTTPConfig.hpp
     PUBLIC
         ${CMAKE_CURRENT_SOURCE_DIR}/include/extreme_carpaccio_client/ExtremeCarpaccioClient.hpp
         ${CMAKE_CURRENT_SOURCE_DIR}/include/extreme_carpaccio_client/config.hpp
-		${CMAKE_CURRENT_SOURCE_DIR}/include/extreme_carpaccio_client/Order.hpp
 )
 
 target_include_directories(Extreme-Carpaccio-Client-shared
@@ -64,7 +40,13 @@ target_include_directories(Extreme-Carpaccio-Client-shared
         ${CMAKE_CURRENT_SOURCE_DIR}/include
 )
 
-target_link_libraries(Extreme-Carpaccio-Client-shared PUBLIC Boost::boost PRIVATE nlohmann_json)
+target_link_libraries(Extreme-Carpaccio-Client-shared
+    PRIVATE
+        nlohmann_json
+    PUBLIC 
+        Boost::boost
+        Extreme-Carpaccio-Order-Management-shared
+)
 
 # Client executable
 add_executable(Extreme-Carpaccio-Client)
