+++ clients/cpp/OrderManagement/includes/extreme_carpaccio/order_management/config.hpp
@@ -1,0 +1,14 @@
+#ifndef EXTREME_CARPACCIO_ORDER_MANAGEMENT_CONFIG
+#define EXTREME_CARPACCIO_ORDER_MANAGEMENT_CONFIG
+
+#ifdef _MSC_VER
+#ifdef EXTREME_CARPACCIO_ORDER_MANAGEMENT_EXPORTS
+#define EXTREME_CARPACCIO_ORDER_MANAGEMENT_API __declspec(dllexport)
+#else
+#define EXTREME_CARPACCIO_ORDER_MANAGEMENT_API __declspec(dllimport)
+#endif
+#else
+#define EXTREME_CARPACCIO_ORDER_MANAGEMENT_API
+#endif
+
+#endif // EXTREME_CARPACCIO_ORDER_MANAGEMENT_CONFIG
